separated request log formatter from logger
