+++ src/main/scala/nl.knaw.dans.easy.deposit/logging/readme.md
@@ -1,0 +1,1 @@
+TODO: candidate for dans-scala-lib
